working version of wrapper daqmxni
Wrapper daqmxni as been implemented from nidaqmx library. Simple counter works. Other channels must be tested.
Must add a "timer" to the counter.
