+++ plugin_info.toml
@@ -11,7 +11,7 @@ license = 'MIT'
 
 [plugin-install]
 #packages required for your plugin:
-packages-required = ['pydaqmx','nidaqmx', 'pymodaq>4.0.1']
+packages-required = ['pydaqmx','nidaqmx','pymodaq>4.0.1']
 ##
 
 [features]  # defines the plugin features contained into this plugin
