+++ src/pymodaq_plugins_daqmx/daq_viewer_plugins/plugins_0D/daq_0Dviewer_DAQmx_counter.py
@@ -4,11 +4,8 @@ from pymodaq.utils.data import DataWithAxes, DataToExport, DataSource
 from pymodaq.control_modules.viewer_utility_classes import DAQ_Viewer_base, comon_parameters, main
 from pymodaq.utils.parameter import Parameter
 import nidaqmx
-from pymodaq_plugins_daqmx.hardware.national_instruments.daqmx import DAQmx, \
-    Edge, ClockSettings, Counter, ClockCounter,  TriggerSettings
 
-from PyDAQmx import DAQmx_Val_ContSamps
-
+from pymodaq_plugins_daqmx.hardware.national_instruments.daqmxni import DAQmx, Edge, Counter
 class DAQ_0DViewer_DAQmx_counter(DAQ_Viewer_base):
     """
     Plugin for a 0D PL counter, based on a NI card.
@@ -18,10 +15,6 @@ class DAQ_0DViewer_DAQmx_counter(DAQ_Viewer_base):
          "type": "list", "limits": DAQmx.get_NIDAQ_channels(source_type="Counter")},
         {"title": "Acq. Time (s):", "name": "acq_time",
          "type": "float", "value": 1., "default": 1.},
-        {'title': 'Clock channel:', 'name': 'clock_channel', 'type': 'list',
-         'limits': DAQmx.get_NIDAQ_channels(source_type='Counter')},
-        {"title": "Clk Trig:", "name": "trig",
-         "type": "list", "limits": DAQmx.getTriggeringSources()}
         ]
 
     def ini_attributes(self):
@@ -58,7 +51,7 @@ class DAQ_0DViewer_DAQmx_counter(DAQ_Viewer_base):
         initialized: bool
             False if initialization failed otherwise True
         """
-        self.controller = {"clock": DAQmx(), "counter": DAQmx()}
+        self.controller = DAQmx() #{"clock": DAQmx(), "counter": DAQmx()}
         try:
             self.update_tasks()
             initialized = True
@@ -68,20 +61,15 @@ class DAQ_0DViewer_DAQmx_counter(DAQ_Viewer_base):
             print(e)
             initialized = False
             info = "Error"
-            
-        #self.dte_signal_temp.emit(DataToExport(name='Counts',
-        #                                       data=[DataWithAxes(name='Counts', data=[np.array([0])],
-        #                                       source=DataSource['raw'],
-        #                                       dim='Data0D', labels=['counts (Hz)'])]))
-        
+
         return info, initialized
-    
+
     def close(self):
         """Terminate the communication protocol"""
         pass
         #self.controller["clock"].close()
         #self.controller["counter"].close()
-        
+
     def grab_data(self, Naverage=1, **kwargs):
         """Start a grab from the detector
 
@@ -92,33 +80,34 @@ class DAQ_0DViewer_DAQmx_counter(DAQ_Viewer_base):
         kwargs: dict
             others optionals arguments
         """
-        update = False  # to decide if we do the initial set up or not
+        #update = False  # to decide if we do the initial set up or not
 
 
-        if 'live' in kwargs:
-            if kwargs['live'] != self.live:
-                update = True
-            self.live = kwargs['live']
-            """
-        if 'live' in kwargs:
-            if kwargs['live'] == self.live:
-                update = False  # we are already live
-            self.live = kwargs['live']
-            """
+        #if 'live' in kwargs:
+        #    if kwargs['live'] != self.live:
+        #        update = True
+        #    self.live = kwargs['live']
+        #    """
+        #if 'live' in kwargs:
+        #    if kwargs['live'] == self.live:
+        #        update = False  # we are already live
+        #    self.live = kwargs['live']
+        #    """
 
-        if update:
-            self.update_tasks()
-            self.controller["clock"].start()
-        
-        read_data = self.controller["counter"].readCounter(1, counting_time=self.counting_time)
-        data = read_data/self.counting_time  # convert to cts/s
+        #if update:
+        #    self.update_tasks()
+        #    self.controller["clock"].start()
+
+        read_data = 32#self.controller.readCounter()#, counting_time=self.counting_time)
+        data = read_data #/self.counting_time  # convert to cts/s
         self.dte_signal.emit(DataToExport(name='Counts',
-                                          data=[DataWithAxes(name='Counts', data=[data],
+                                          data=[DataWithAxes(name='Counts', data=[np.array([data])],
                                                              source=DataSource['raw'],
                                                              dim='Data0D', labels=['Counts (Hz)'])]))
 
     def stop(self):
         """Stop the current grab hardware wise if necessary"""
+        self.task.stop()
         self.close()
         self.emit_status(ThreadCommand('Update_Status', ['Acquisition stopped.']))
         return ''
@@ -126,24 +115,22 @@ class DAQ_0DViewer_DAQmx_counter(DAQ_Viewer_base):
     def update_tasks(self):
         """Set up the counting tasks in the NI card."""
         # Create channels
-        self.clock_channel = ClockCounter(1/self.settings.child("acq_time").value(),
-                                          name=self.settings.child("trig").value(),
-                                          source="Counter")
+
         self.counter_channel = Counter(name=self.settings.child("counter_channel").value(),
                                        source="Counter", edge=Edge.names()[0])
-
-        self.controller["clock"].update_task(channels=[self.clock_channel],
-                                             clock_settings=ClockSettings(),
-                                             trigger_settings=TriggerSettings())
-        self.controller["clock"].task.CfgImplicitTiming(DAQmx_Val_ContSamps, 1)
-        
-        self.controller["counter"].update_task(channels=[self.counter_channel],
-                                               clock_settings=ClockSettings(),
-                                               trigger_settings=TriggerSettings())
+        #
+        #self.controller["clock"].update_task(channels=[self.clock_channel],
+        #                                     clock_settings=ClockSettings(),
+        #                                     trigger_settings=TriggerSettings())
+        #self.controller["clock"].task.CfgImplicitTiming(DAQmx_Val_ContSamps, 1)
+        #
+        #self.controller["counter"].update_task(channels=[self.counter_channel],
+        #                                       clock_settings=ClockSettings(),
+        #                                       trigger_settings=TriggerSettings())
 
         # connect the clock to the counter
-        self.controller["counter"].task.SetSampClkSrc("/" + self.clock_channel.name + "InternalOutput")
+        #self.controller["counter"].task.SetSampClkSrc("/" + self.clock_channel.name + "InternalOutput")
 
-        
+
 if __name__ == '__main__':
     main(__file__)
